added payment success to prescription renewal tests, added fields to prescription renewal, added security checks for payment success

+++ src/main/kotlin/fi/metatavu/ikioma/rest/V1ApiImpl.kt
@@ -8,7 +8,7 @@ import fi.metatavu.ikioma.email.api.spec.model.PaymentStatus
 import fi.metatavu.ikioma.email.api.spec.model.PrescriptionRenewal
 import fi.metatavu.ikioma.rest.translate.PrescriptionRenewalTranslator
 import org.slf4j.Logger
-import java.net.URI
+import java.lang.IllegalArgumentException
 import java.util.*
 import javax.annotation.security.RolesAllowed
 import javax.enterprise.context.RequestScoped
@@ -43,6 +43,7 @@ class V1ApiImpl : V1Api, AbstractApi() {
         TODO("Not yet implemented")
     }
 
+    @RolesAllowed(value = [UserRole.PATIENT.name])
     override fun checkoutFinlandSuccess(
         checkoutAccount: Int,
         checkoutAlgorithm: String,
@@ -54,14 +55,36 @@ class V1ApiImpl : V1Api, AbstractApi() {
         checkoutProvider: String,
         signature: String
     ): Response {
-        //todo is there token
-        //todo verify signature?
-        val prescriptionRenewal = prescriptionController.findPrescriptionRenewalByTransactionId(transactionId = checkoutTransactionId)
-        prescriptionRenewal ?: return createNoContent()
+        val userId = loggedUserId ?: return createUnauthorized("Unauthorized")
+        val refNo: UUID?
+        try {
+            refNo = UUID.fromString(checkoutReference)
+        } catch (e: IllegalArgumentException) {
+            return createBadRequest("Invalid reference no")
+        }
+
+        val prescriptionRenewal = prescriptionController.findPrescriptionRenewalByReference(reference = refNo)
+        prescriptionRenewal ?: return createNotFound()
+
+        if (prescriptionRenewal.stamp != UUID.fromString(checkoutStamp) || prescriptionRenewal.transactionId != checkoutTransactionId) {
+            return createForbidden("Payment information does not match")
+        }
+
+        if (!paymentController.verifyPayment(signature, mapOf(
+                Pair("checkout-account" ,checkoutAccount),
+                Pair("checkout-algorithm", checkoutAlgorithm),
+                Pair("checkout-amount", checkoutAmount),
+                Pair("checkout-stamp", checkoutStamp),
+                Pair("checkout-reference", checkoutReference),
+                Pair("checkout-transaction-id", checkoutTransactionId),
+                Pair("checkout-status", checkoutStatus),
+                Pair("checkout-provider", checkoutProvider)))) {
+            return createForbidden("Bad signature")
+        }
 
         prescriptionController.updatePrescriptionRenewalStatus(prescriptionRenewal, PaymentStatus.PAID)
         //send email
-        //emailController.sendPrescriptionRenewalSuccess(prescriptionRenewal)
+       // emailController.sendPrescriptionRenewalSuccess(prescriptionRenewal, userId, "ssn")
 
         prescriptionController.deletePrescriptionRenewal(prescriptionRenewal)
         return createOk()
@@ -75,11 +98,14 @@ class V1ApiImpl : V1Api, AbstractApi() {
         paymentData ?: return createInternalServerError("Failed to create payment")
 
         val newPrescriptionRenewal = prescriptionController.createPrescriptionRenewal(
+            id = paymentData.id,
+            stamp = paymentData.stamp,
             prescriptions = prescriptionRenewal.prescriptions,
             practitionerUserId = prescriptionRenewal.practitionerUserId,
             paymentStatus = prescriptionRenewal.status,
             paymentUrl = paymentData.paymentUrl,
             transactionId = paymentData.transactionId,
+            checkoutAccount = paymentData.checkoutAccount,
             userId = userId
         )
